% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{boxplot_byDRG}
\alias{boxplot_byDRG}
\title{boxplot_byDRG}
\usage{
boxplot_byDRG(data, payment_type, x_name)
}
\arguments{
\item{data}{is the input dataset}

\item{payment_type}{is the type of payment chose}

\item{x_name}{is the name of the variable in x-axis}
}
\value{
a boxplot with DRX markers of \code{payment_type}
}
\description{
Function returns a box plot of payments by DRG code, where the payment type is choosen.
}
\examples{
boxplot_byDRG(DRG,DRG$`Average Medicare Payments`, "Average Medicare Payments")
}
